Update Timeline test cases
