+++ e2e-tests/tests/teacher.spec.ts
@@ -20,11 +20,11 @@ import {
 import {
   downloadCSVGradeTemplate,
   downloadExcelGradeTemplate,
-  filterGradesTable,
   importGradesWithFile,
   importGradesWithText,
 } from './common/grades';
 import {addCoursePartTaskFail, addCoursePartTaskSuccess} from './common/task';
+import {searchTimeline, viewTimeline} from './common/timeline';
 import {aPlusToken} from './common/token';
 import {logOut} from './common/user';
 import {setupDb} from './helper';
@@ -124,9 +124,15 @@ test.describe('Test courses as teacher', () => {
     });
   });
 
-  test.describe('Grades table', () => {
-    test('Filter results', async ({page}) => {
-      await filterGradesTable(page);
+  test.describe('Timeline', () => {
+    test('Teacher can view timeline', async ({page}) => {
+      await checkCourse(page);
+      await viewTimeline(page);
+    });
+
+    test('Teacher can search in timeline', async ({page}) => {
+      await checkCourse(page);
+      await searchTimeline(page, 'NonExistentStudent');
     });
   });
